@@ -5,9 +5,10 @@ import 'package:meta/meta.dart';
 
 @immutable
 class Task extends Equatable {
+  // TODO: need id property
   final bool isDone;
-  final DateTime dateCompleted;
-  final DateTime dueDate;
+  final DateTime? dateCompleted;
+  final DateTime? dueDate;
   final String title;
   final String description;
   final List<Subtask> subtasks;
@@ -26,53 +27,62 @@ class Task extends Equatable {
   int compareByDueDate(Task other) {
     if (null == this.dueDate) {
       if (null == other.dueDate) {
+        // both due dates are null - use priority instead.
         return this.priority.index.compareTo(other.priority.index);
       }
-      return -1;
+      return -1; // this (dueDate) is null, other is not
+    } else {
+      if (null == other.dueDate) {
+        return 1; // this (dueDate) is not null, other is
+      } else {
+        return this
+            .dueDate!
+            .compareTo(other.dueDate!); // neither dueDate is null
+      }
     }
-    if (null == other.dueDate) return 1;
-    return this.dueDate.compareTo(other.dueDate);
   }
 
   bool get isOverDue {
+    if (null == dueDate) return false;
     var today = DateTime.now();
     today = DateTime(today.year, today.month, today.day, 0, 0, 0);
-    if (dueDate.isBefore(today)) return true;
-    return false;
+    return dueDate!.isBefore(today);
   }
 
   bool get isDueToday {
     if (dueDate == null) return false;
     var today = DateTime.now();
-    if (dueDate.year == today.year &&
-        dueDate.month == today.month &&
-        dueDate.day == today.day) return true;
+    if (dueDate!.year == today.year &&
+        dueDate!.month == today.month &&
+        dueDate!.day == today.day) return true;
     return false;
   }
 
   bool get isDueTomorrow {
     if (dueDate == null) return false;
     var tomorrow = DateTime.now().add(Duration(days: 1));
-    if (dueDate.year == tomorrow.year &&
-        dueDate.month == tomorrow.month &&
-        dueDate.day == tomorrow.day) return true;
+    if (dueDate!.year == tomorrow.year &&
+        dueDate!.month == tomorrow.month &&
+        dueDate!.day == tomorrow.day) return true;
     return false;
   }
 
   bool get isDueThisWeek {
     if (dueDate == null) return false;
     var today = DateTime.now();
-    int daysToNextWeek = 7 - today.weekday;
+    int daysToNextWeek = 8 - today.weekday;
     var firstOfNextWeek = today.add(Duration(days: daysToNextWeek));
     firstOfNextWeek = DateTime(
       firstOfNextWeek.year,
       firstOfNextWeek.month,
       firstOfNextWeek.day,
     );
-    return dueDate.isBefore(firstOfNextWeek);
+    return dueDate!.isBefore(firstOfNextWeek);
   }
 
   bool get isDueThisWeekOrUndated {
+    // Weekdays:
+    // 1=mon,2=tue,3=wed,4=thu,5=fri,6=sat,7=sun
     if (dueDate == null) return true;
     var today = DateTime.now();
     int daysToNextWeek = 8 - today.weekday;
@@ -82,7 +92,7 @@ class Task extends Equatable {
       firstOfNextWeek.month,
       firstOfNextWeek.day,
     );
-    return dueDate.isBefore(firstOfNextWeek);
+    return dueDate!.isBefore(firstOfNextWeek);
   }
 
   bool get isDueAfterThisWeek {
@@ -92,25 +102,31 @@ class Task extends Equatable {
     var lastDayOfWeek = today.add(Duration(days: daysToLastWeekday));
     lastDayOfWeek = DateTime(lastDayOfWeek.year, lastDayOfWeek.month,
         lastDayOfWeek.day, 23, 59, 59, 999);
-    return dueDate.isAfter(lastDayOfWeek);
+    return dueDate!.isAfter(lastDayOfWeek);
   }
 
   String get dueDateString {
-    if (null == dueDate) return "Not set";
-    if (isDueToday) return "Today";
-    if (isDueTomorrow) return "Tomorrow";
-    var today = DateTime.now();
-    if (dueDate.isAfter(today) &&
-        dueDate.isBefore(today.add(Duration(days: 7)))) {
-      if (1 == dueDate.weekday) return "Monday";
-      if (2 == dueDate.weekday) return "Tuesday";
-      if (3 == dueDate.weekday) return "Wednesday";
-      if (4 == dueDate.weekday) return "Thursday";
-      if (5 == dueDate.weekday) return "Friday";
-      if (6 == dueDate.weekday) return "Saturday";
-      if (7 == dueDate.weekday) return "Sunday";
+    if (null == dueDate) {
+      return "Not set";
+    } else {
+      if (isDueToday) return "Today";
+      if (isDueTomorrow) return "Tomorrow";
+      var today = DateTime.now();
+      today = DateTime(today.year, today.month, today.day, 0, 0, 0);
+      if (dueDate!.isAfter(today) &&
+          dueDate!.isBefore(today.add(Duration(days: 7)))) {
+        // TODO: should be `Duration(days: 8)`? since `isBefore.
+        // TODO: Write unit tests
+        if (1 == dueDate!.weekday) return "Monday";
+        if (2 == dueDate!.weekday) return "Tuesday";
+        if (3 == dueDate!.weekday) return "Wednesday";
+        if (4 == dueDate!.weekday) return "Thursday";
+        if (5 == dueDate!.weekday) return "Friday";
+        if (6 == dueDate!.weekday) return "Saturday";
+        if (7 == dueDate!.weekday) return "Sunday";
+      }
+      return "${dueDate!.month}/${dueDate!.day}/${dueDate!.year}";
     }
-    return "${dueDate.month}/${dueDate.day}/${dueDate.year}";
   }
 
   String get priorityString {
@@ -128,6 +144,6 @@ class Task extends Equatable {
       subtasks.where((Subtask subtask) => subtask.isDone).length;
 
   @override
-  List<Object> get props =>
+  List<Object?> get props =>
       [isDone, dateCompleted, dueDate, title, description, subtasks, priority];
 }
