@@ -25,9 +25,14 @@ class Data {
       .where((Task task) => task.isDueAfterThisWeek)
       .toList();
 
-  Data({this.habits, this.tasks, this.weeklyReport, this.userHistory});
+  Data(
+      {required this.habits,
+      required this.tasks,
+      required this.weeklyReport,
+      required this.userHistory});
 }
 
+final today = DateTime.now();
 final data = Data(
   habits: [
     Habit(title: "Brush your teeth"),
@@ -89,10 +94,14 @@ final data = Data(
     ),
   ],
   weeklyReport: WeeklyReport(
+    weekBegins: DateTime(today.year, today.month, today.day, 0, 0, 0)
+        .subtract(Duration(days: today.weekday)),
     tasksCompleted: 5,
     tasksDue: 7,
     tasksDoneOnTimeStreak: 10,
+    tasksCreated: 0,
     dailyHabitStreak: 36,
+    weeklyHabitStreak: 0,
     tasksDoneInWeekNewBest: false,
     onTimeNewBestStreak: false,
     habitsNewBestStreak: true,
