@@ -3,32 +3,33 @@ import 'package:meta/meta.dart';
 
 @immutable
 class WeeklyReport extends Equatable {
-  DateTime weekBegins;
-  int tasksCreated;
-  int tasksCompleted;
-  int tasksDue;
-  int tasksDoneOnTimeStreak;
-  int dailyHabitStreak;
-  int weeklyHabitStreak;
-  bool tasksDoneInWeekNewBest;
-  bool onTimeNewBestStreak;
-  bool habitsNewBestStreak;
-  List<bool> habitsCompletedDays;
+  final DateTime weekBegins;
+  final int tasksCreated;
+  final int tasksCompleted;
+  final int tasksDue;
+  final int tasksDoneOnTimeStreak;
+  final int dailyHabitStreak;
+  final int weeklyHabitStreak;
+  final bool tasksDoneInWeekNewBest;
+  final bool onTimeNewBestStreak;
+  final bool habitsNewBestStreak;
+  final List<bool> habitsCompletedDays;
 
   int get numDayshabitsDone =>
       habitsCompletedDays.where((bool completed) => completed).length;
 
   WeeklyReport({
-    this.weekBegins,
-    this.tasksCreated,
-    this.tasksCompleted,
-    this.tasksDue,
-    this.tasksDoneOnTimeStreak,
-    this.dailyHabitStreak,
-    this.tasksDoneInWeekNewBest,
-    this.onTimeNewBestStreak,
-    this.habitsNewBestStreak,
-    this.habitsCompletedDays,
+    required this.weekBegins,
+    required this.tasksCreated,
+    required this.tasksCompleted,
+    required this.tasksDue,
+    required this.tasksDoneOnTimeStreak,
+    required this.dailyHabitStreak,
+    required this.weeklyHabitStreak,
+    required this.tasksDoneInWeekNewBest,
+    required this.onTimeNewBestStreak,
+    required this.habitsNewBestStreak,
+    required this.habitsCompletedDays,
   });
 
   @override
